# Small performance improvement to `process_item`