--- a/sub-packages/bionemo-geneformer/pyproject.toml
+++ b/sub-packages/bionemo-geneformer/pyproject.toml
@@ -14,7 +14,6 @@ dependencies = [
     # bionemo sub-packages
     'bionemo-core',
     'bionemo-llm',
-    'bionemo-testing',  # needed for getting the tokenizer from NGC
     # external
     'cellxgene_census',
 ]