--- a/sub-packages/bionemo-geneformer/scripts/README.md
+++ b/sub-packages/bionemo-geneformer/scripts/README.md
@@ -1,0 +1,3 @@
+# WARNING
+This folder contains one-off eval scripts that may not run and are not actively tested or kept up to date.
+Also these scripts may depend on `bionemo-testing` which is generally not allowed.