--- a/sub-packages/bionemo-geneformer/src/bionemo/geneformer/data/singlecell/dataset.py
+++ b/sub-packages/bionemo-geneformer/src/bionemo/geneformer/data/singlecell/dataset.py
@@ -285,15 +285,17 @@ def process_item(  # noqa: D417
 
     gene_names = feature_ids[gene_idxs]
 
-    genes, token_ids, medians = _gather_medians(gene_names, gene_data, normalize, tokenizer.vocab, gene_median)
+    gene_expression, token_ids, medians = _gather_medians(
+        gene_names, gene_data, normalize, tokenizer.vocab, gene_median
+    )
 
     if normalize:
         # re-order according to expression median normalized rank. descending order.
 
-        genes = genes / genes.sum() * target_sum
-        genes = genes / medians.astype(float)
-        idxs = np.argsort(-genes)  # sort in descending order so that the 0th position is the highest value.
-        genes = genes[idxs]
+        gene_expression = gene_expression / gene_expression.sum() * target_sum
+        gene_expression = gene_expression / medians.astype(float)
+        idxs = np.argsort(-gene_expression)  # sort in descending order so that the 0th position is the highest value.
+        gene_expression = gene_expression[idxs]
         token_ids = token_ids[idxs]
 
     # - select max_len subset, set sample to false so it doesnt permute the already rank ordered expression values.