--- a/tach.toml
+++ b/tach.toml
@@ -54,7 +54,6 @@ path = "bionemo.geneformer"
 depends_on = [
     { path = "bionemo.core" },
     { path = "bionemo.llm" },
-    { path = "bionemo.testing" },  # needed for the inference script to get the tokenizer reliably.
 ]
 
 [[modules]]