## Summary
Run `python -m bionemo.geneformer.data.singlecell.dataset`
### Baseline:
Processed 31208 rows in 47.10491418838501 seconds
Processed 31208 rows in 47.388004779815674 seconds
### After vectorization: 
Processed 31208 rows in 44.2215359210968 seconds
Processed 31208 rows in 44.54389500617981 seconds